requirements.txt
github3.py>=1.0.0a2
